+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+<!-- ## Unreleased -->
+
+## [0.14.10] - 2025-01-28
 
 ### Fixed
 
@@ -643,7 +645,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release.
 
-[unreleased]: https://github.com/google/wireit/compare/v0.14.9...HEAD
+[unreleased]: https://github.com/google/wireit/compare/v0.14.10...HEAD
+[0.14.10]: https://github.com/google/wireit/compare/v0.14.9...v0.14.10
 [0.14.9]: https://github.com/google/wireit/compare/v0.14.8...v0.14.9
 [0.14.8]: https://github.com/google/wireit/compare/v0.14.7...v0.14.8
 [0.14.7]: https://github.com/google/wireit/compare/v0.14.6...v0.14.7